requirements.txt
langchain
openai
pypdf
python-dotenv
streamlit
